requirements.txt
qwen_vl_utils
accelerate
torch-snippets
pandas
numpy
streamlit
torch>=1.7.0
transformers
Pillow
requests
loguru
torchvision

#python-bidi
#pyclipper